Improve HSElementHooker
watchElement now accepts optional "greedy" argument. This makes the hook trigger the supplied callback on DOM change even if the value itself did not change.

Using greedy can result in a lot of callback triggers if the watched element is updated frequently in the DOM. To counter this, there is now innate 200ms throttling implemented in all watchers. In other words, even if there are changes in the dom and watch is set to greedy, it will only ever trigger callback every 200ms max.