Make HSModuleManager static
Eventually all modules should extend from HSModule. The only reason this isn't so is that it's kind of annoying to use HSModuleManager to get module instances because you'd have to keep passing HSModuleManager around and use getModule every time.

If HSModuleManager is made static, there is no need to pass it around.